Move bootstrap functionality to a new package. (#3176)
